bundling: change configuration from grails.plugins.pageResources.bundlePageResources to grails.plugins.pageResources.defaultBundle
This has a couple advantages.  First, the naming is now more similar to how things are named in the resource plugin, which should make it
easier for users to remember and understand.  Second, the behavior has been enhanced so that now, instead of only supporting true/false,
it also supports String values, in exactly the same way as configuring defaultBundle in a resource module manually.  The user experience
should now be that setting a value in this config key has the same effect as if they had included an equivalent defaultBundle entry in each
page module manually.